fix silly things

+++ RaikesSimplexSolver/UnitTests/SolverTest.cs
@@ -161,10 +161,10 @@ namespace UnitTests
             String mat = mm.ToString();
             mat = mat.Replace("  ", "\t");
             System.Diagnostics.Debug.WriteLine(mat);
-            var subM = target.findBasicMatrix(mm, new[] { 1.0, 2.0 });
+            var subM = target.findBasicMatrix(mm, new[] { 1, 2 });
             System.Diagnostics.Debug.WriteLine(subM);
 
-            double[] array = target.basicColumnIndecies(model);
+            int[] array = target.basicColumnIndecies(model);
             System.Diagnostics.Debug.WriteLine(array[0] + " 1: " + array[1] + " 2: " + array[2] + " 3: " + array[3]);
             Matrix<double> mmm = target.convertToMatrix(model);
             System.Diagnostics.Debug.WriteLine(target.findBasicMatrix(mmm, array));
